cleaned up deleteafter
