+++ Railgun/Commands/JoinLeave/JoinLeave.cs
@@ -27,23 +27,6 @@ namespace Railgun.Commands.JoinLeave
 			return ReplyAsync($"Join/Leave Notifications is now {Format.Bold((data.ChannelId == 0 ? "Enabled & Set" : "Set"))} to this channel.");
 		}
 
-		[Command("deleteafter"), BotPerms(ChannelPermission.ManageMessages)]
-		public Task DeleteAfterAsync(int minutes = 0)
-		{
-			if (minutes < 0) return ReplyAsync("Minutes can not be less than 0.");
-
-			var profile = Context.Database.ServerProfiles.GetOrCreateData(Context.Guild.Id);
-            var data = profile.JoinLeave;
-
-			if (minutes == 0 && data.DeleteAfterMinutes == 0) {
-				return ReplyAsync("Already set to not delete Join/Leave notifications.");
-			} else if (minutes == 0 && data.DeleteAfterMinutes != 0) {
-				data.DeleteAfterMinutes = 0;
-				return ReplyAsync("No longer deleting Join/Leave notifications.");
-			}
-
-			data.DeleteAfterMinutes = minutes;
-			return ReplyAsync($"Join/Leave notifications will now be deleted after {minutes} minutes.");
-		}
+		
 	}
 }
